+++ src/config/chain/initia.ts
@@ -3,7 +3,7 @@ import { wallets as keplrWallets } from "@cosmos-kit/keplr";
 import type { ChainConfigs } from "./types";
 
 export const INITIA_CHAIN_CONFIGS: ChainConfigs = {
-  "stone-9": {
+  "stone-10": {
     chain: "initia",
     registryChainName: "initiatestnet",
     prettyName: "Initia Testnet",