+++ src/lib/services/module.ts
@@ -6,6 +6,8 @@ import type {
   ResponseModules,
   InternalModule,
   HexAddr,
+  Option,
+  SnakeToCamelCaseNested,
 } from "lib/types";
 import { snakeToCamel } from "lib/utils";
 
@@ -56,14 +58,26 @@ interface ModuleVerificationReturn {
   chain_id: string;
 }
 
+// TODO: Figure out how to correctly infer NominalType intersection
+export interface ModuleVerificationInternal
+  extends Omit<
+    SnakeToCamelCaseNested<ModuleVerificationReturn>,
+    "moduleAddress"
+  > {
+  moduleAddress: HexAddr;
+}
+
 export const getModuleVerificationStatus = async (
   address: MoveAccountAddr,
   moduleName: string
-): Promise<boolean> =>
+): Promise<Option<ModuleVerificationInternal>> =>
   // TODO: move url to base api route? wait for celatone api implementation?
   axios
     .get<ModuleVerificationReturn>(
       `https://stone-compiler.initia.tech/contracts/${address}/${moduleName}`
     )
-    .then(() => true)
-    .catch(() => false);
+    .then(({ data }) => ({
+      ...snakeToCamel(data),
+      moduleAddress: data.module_address,
+    }))
+    .catch(() => undefined);