+++ src/lib/types/abi.ts
@@ -1,4 +1,3 @@
-import type { HexAddr } from "./addrs";
 import type { SnakeToCamelCaseNested } from "./converter";
 
 export enum UpgradePolicy {
@@ -27,8 +26,9 @@ interface ABIFunction {
 }
 
 /* response */
+// TODO: change address type to HexAddr after figuring out how to correctly infer NominalType intersection
 export interface ResponseModule {
-  address: HexAddr;
+  address: string;
   module_name: string;
   abi: string;
   raw_bytes: string;
@@ -45,8 +45,9 @@ export interface ModulePagination {
   total: string;
 }
 
+// TODO: change address type to HexAddr after figuring out how to correctly infer NominalType intersection
 export interface ResponseABI {
-  address: HexAddr;
+  address: string;
   name: string;
   friends: string[];
   exposed_functions: ExposedFunction[];